+++ Controller/SettingsController.php
@@ -1,16 +1,16 @@
 <?php
 
 /**
- * This file is part of the DmishhSettingsBundle package.
- * (c) 2013 Dmitriy Scherbina <http://dmishh.com>
+ * This file is part of the MharySettingsBundle package.
+ * (c) 2013 Dmitriy Scherbina <http://mhary.com>
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Dmishh\SettingsBundle\Controller;
+namespace Mhary\SettingsBundle\Controller;
 
-use Dmishh\SettingsBundle\Entity\SettingsOwnerInterface;
-use Dmishh\SettingsBundle\Form\Type\SettingsType;
+use Mhary\SettingsBundle\Entity\SettingsOwnerInterface;
+use Mhary\SettingsBundle\Form\Type\SettingsType;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;