+++ Entity/Setting.php
@@ -1,21 +1,21 @@
 <?php
 
 /**
- * This file is part of the DmishhSettingsBundle package.
+ * This file is part of the MharySettingsBundle package.
  *
- * (c) 2013 Dmitriy Scherbina <http://dmishh.com>
+ * (c) 2013 Dmitriy Scherbina <http://mhary.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Dmishh\SettingsBundle\Entity;
+namespace Mhary\SettingsBundle\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
 
 /**
  * @ORM\Table(
- *  name="dmishh_settings",
+ *  name="mhary_settings",
  *  indexes={@ORM\Index(name="name_owner_id_idx", columns={"name", "owner_id"})}
  * )
  * @ORM\Entity