+++ Manager/SettingsManagerInterface.php
@@ -1,15 +1,15 @@
 <?php
 
 /**
- * This file is part of the DmishhSettingsBundle package.
- * (c) 2013 Dmitriy Scherbina <http://dmishh.com>
+ * This file is part of the MharySettingsBundle package.
+ * (c) 2013 Dmitriy Scherbina <http://mhary.com>
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Dmishh\SettingsBundle\Manager;
+namespace Mhary\SettingsBundle\Manager;
 
-use Dmishh\SettingsBundle\Entity\SettingsOwnerInterface;
+use Mhary\SettingsBundle\Entity\SettingsOwnerInterface;
 
 interface SettingsManagerInterface
 {