+++ Serializer/JsonSerializer.php
@@ -1,15 +1,15 @@
 <?php
 
 /**
- * This file is part of the DmishhSettingsBundle package.
+ * This file is part of the MharySettingsBundle package.
  *
- * (c) 2013 Dmitriy Scherbina <http://dmishh.com>
+ * (c) 2013 Dmitriy Scherbina <http://mhary.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Dmishh\SettingsBundle\Serializer;
+namespace Mhary\SettingsBundle\Serializer;
 
 class JsonSerializer implements SerializerInterface
 {