+++ Serializer/SerializerFactory.php
@@ -1,17 +1,17 @@
 <?php
 
 /**
- * This file is part of the DmishhSettingsBundle package.
+ * This file is part of the MharySettingsBundle package.
  *
- * (c) 2013 Dmitriy Scherbina <http://dmishh.com>
+ * (c) 2013 Dmitriy Scherbina <http://mhary.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Dmishh\SettingsBundle\Serializer;
+namespace Mhary\SettingsBundle\Serializer;
 
-use Dmishh\SettingsBundle\Exception\UnknownSerializerException;
+use Mhary\SettingsBundle\Exception\UnknownSerializerException;
 use Symfony\Component\DependencyInjection\Container;
 
 class SerializerFactory
@@ -19,13 +19,13 @@ class SerializerFactory
     /**
      * @param string $name short name of serializer (ex.: php) or full class name
      *
-     * @throws \Dmishh\SettingsBundle\Exception\UnknownSerializerException
+     * @throws \Mhary\SettingsBundle\Exception\UnknownSerializerException
      *
      * @return SerializerInterface
      */
     public static function create($name)
     {
-        $serializerClass = 'Dmishh\\SettingsBundle\\Serializer\\' . Container::camelize($name) . 'Serializer';
+        $serializerClass = 'Mhary\\SettingsBundle\\Serializer\\' . Container::camelize($name) . 'Serializer';
 
         if (class_exists($serializerClass)) {
             return new $serializerClass();