+++ Tests/Resources/app/AppKernel.php
@@ -13,7 +13,7 @@ class AppKernel extends Kernel
         $bundles = [
             new \Symfony\Bundle\FrameworkBundle\FrameworkBundle(),
             new \Doctrine\Bundle\DoctrineBundle\DoctrineBundle(),
-            new \Dmishh\SettingsBundle\DmishhSettingsBundle(),
+            new \Mhary\SettingsBundle\MharySettingsBundle(),
         ];
 
         return $bundles;